shacl/requirements.txt
pyshacl
pyyaml